Properly resolve all new contents in PackImpl.applyTo() using a PreorderDiffWalk
When a Pack being sent to a remote through PackImpl.applyTo(...), the
set of objects to transfer is computed by a PreorderDiffWalk against
each commit and each of its parents, with ObjectId deduplication.

Before the fix for all the buckets to be notified to a
PreorderDiffWalk.Consumer, some bucket trees were missing and hence it
was not possible to properly implement the missing contents resolution
this way.

Signed-off-by: Gabriel Roldan <groldan@boundlessgeo.com>