Allow DataStore to receive Context instead of Repository
DataStores may be created as short lived objects out
of a repository or context (possibly a transaction).

This patch allows to instruct the DataStore to not close
the repository and to receive the context as constructor
argument in case it needs to be used as a short lived
object inside an already existing transaction.