Fix a bug that prevented screenmap filter being applied with some valid filters
The geotools renderer often sends a filter of the type BBOX OR BBOX,
which wasn't being added to the pre-filter and hence the screenmap
filter coulnd't be applied, leading to a full tree traversal for "zoomed
out" map requests.

This path fixes PrePostFilterSplitter to replace the property name
in BBOX filters by the @bounds meta property, as stated in the class'
javadocs, making the PreFilter properly handling supported filters and
hence also avoiding the need for a separate bounds filter set to the
diff traversal in FeatureReaderBuilder.

Signed-off-by: Gabriel Roldan <groldan@boundlessgeo.com>