Allow to perform diffs against two quad trees.
PreorderDiffWalk relied solely on CanonicalNodeOrder to perform
the diff walk, which prevented creating a stream of diff events
out of two quad trees.

In order for the DataStore to take advantage of the spatial index
when used as a "diff feature source", PreorderDiffWalk is now given
a NodeOrder object, which is either the CanonicalNodeOrder as usual
or one created by QuadTreeClusteringStrategy for spatial indices.

Signed-off-by: Gabriel Roldan <groldan@boundlessgeo.com>