Add postgres backend commands to create DDL and upgrade database
CreateDDL creates the DDL script that's used to initialize a
PostgreSQL geogig database.

PGDatabaseUpgrade upgrades a geogig database from geogig < 1.2.1,
by adding the dstindex column to the geogig_graph_edge table and
running rebuild-graph on all its repositories. This is necessary
to ensure the proper order of parent commit ids when traversing
the commit graph. Otherwise merge commits might be reported to have
the left and right source commits swapped.

Signed-off-by: Gabriel Roldan <groldan@boundlessgeo.com>