RefDatabase API improvements
RefDatbase used to work with key/value pairs
where the key was the ref name and the value
either an ObjectId or a String with the target
ref name for symbolic references.

With Ref being a core model object it makes sense
that RefDatabase works with Ref directly.

Some methods where added to handle atomic operations
on several refs, and a new command UpdateRefs introduced
to do so at the command level. Most commands that update
several refs were modified to use UpdateRefs instead
to doing several single-ref changes.

Consequently, UpdateRef and UpdateSymRef were changed
to delegeate to UpdateRefs, so that there's a single
command in charge of updating the RefDatabase objects.

Note setting a "reason for change" property on all three
ref updating commands is temporarily mandatory in order
to catch any missing but important ref update that could
be a target for the reflog, for which an initial no-op
command hook was introduced.

Signed-off-by: Gabriel Roldan <gabriel.roldan@gmail.com>